# internal/api/docs/swagger.yaml
components:
  schemas:
...
    api.apiError:
      properties:
        code:
          enum:
          - invalid_json
          - unsupported_media_type
          - not_logged_in
          - session_expired
          - admin_only
          - missing_profile
          - unknown_profile
          - internal
          - not_found
          - invalid_credentials
          - already_set_up
          - username_required
          - password_length
          - unsupported_language
          - wrong_password
          - name_required
          - unknown_avatar
          - profile_limit
          - last_profile
          - username_taken
          - delete_self
          - invalid_id
          - invalid_request
          - episode_required
          - tmdb_unreachable
          - invalid_link
          - link_dead
          - invalid_progress
          - flaresolverr_unreachable
          - device_code_expired
          - device_code_invalid
          - release_unreachable
          - too_many_attempts
          type: string
        error:
          type: string
      required:
      - code
      - error
      type: object
    api.credentials:
      properties:
        password:
          type: string
        username:
          type: string
      required:
      - password
      - username
      type: object
...
    api.flareSolverrSettings:
      properties:
        url:
          description: URL of the FlareSolverr server, "" when disabled.
          type: string
      required:
      - url
      type: object
...
    api.instanceUpdate:
      properties:
        language:
          enum:
          - en
          - fr
          type: string
      required:
      - language
      type: object
...
    api.newUser:
      properties:
        isAdmin:
          type: boolean
        password:
          type: string
        username:
          type: string
      required:
      - isAdmin
      - password
      - username
      type: object
    api.passwordChange:
      properties:
        current:
          type: string
        new:
          type: string
      required:
      - current
      - new
      type: object
...
    api.profileBody:
      properties:
        avatar:
          type: string
        name:
          type: string
        skipSegments:
          description: SkipSegments shows the "skip intro" and "skip credits" buttons.
          type: boolean
      required:
      - avatar
      - name
      - skipSegments
      type: object
...
    api.setupRequest:
      properties:
        language:
          enum:
          - en
          - fr
          type: string
        password:
          type: string
        username:
          type: string
      required:
      - language
      - password
      - username
      type: object
    api.userUpdate:
      properties:
        maxStreamMbps:
          description: 'MaxStreamMbps caps the throughput of each stream the account
            plays, in Mbit/s (0: unlimited).'
          maximum: 10000
          minimum: 0
          type: integer
      required:
      - maxStreamMbps
      type: object
...
    store.Profile:
      properties:
        avatar:
          type: string
        id:
          type: integer
        name:
          type: string
        skipSegments:
          description: SkipSegments shows the "skip intro" and "skip credits" buttons.
          type: boolean
      required:
      - avatar
      - id
      - name
      - skipSegments
      type: object
...
    store.User:
      properties:
        id:
          type: integer
        isAdmin:
          type: boolean
        maxStreamMbps:
          description: 'MaxStreamMbps caps the throughput of each stream the account
            plays, in Mbit/s (0: unlimited).'
          type: integer
        username:
          type: string
      required:
      - id
      - isAdmin
      - maxStreamMbps
      - username
      type: object
...
  securitySchemes:
    SessionCookie:
      in: header
      name: X-Profile-ID
      type: apiKey
...
info:
  description: Self-hosted multi-source streaming (Go port of vStream).
  title: kinora API
  version: "1.0"
openapi: 3.1.0
paths:
  /admin/flaresolverr:
    get:
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.flareSolverrSettings'
          description: OK
      summary: FlareSolverr server used for sources behind Cloudflare
      tags:
      - admin
    put:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/api.flareSolverrSettings'
              description: FlareSolverr URL
              summary: body
        description: FlareSolverr URL
        required: true
      responses:
        "204":
          description: No Content
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.apiError'
          description: Bad Request
      summary: Set the FlareSolverr server (empty URL disables it), checked before
        saving. Sources behind Cloudflare are queried while it is set.
      tags:
      - admin
  /admin/instance:
    put:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/api.instanceUpdate'
              description: Language
              summary: body
        description: Language
        required: true
      responses:
        "204":
          description: No Content
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.apiError'
          description: Bad Request
      summary: Change the instance language (catalog, UI and sources)
      tags:
      - admin
...
  /admin/users:
    get:
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/store.User'
                type: array
          description: OK
      summary: List accounts
      tags:
      - admin
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/api.newUser'
              description: Account
              summary: body
        description: Account
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/store.User'
          description: Created
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.apiError'
          description: Conflict
      summary: Create an account
      tags:
      - admin
  /admin/users/{id}:
    delete:
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        schema:
          type: integer
      responses:
        "204":
          description: No Content
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.apiError'
          description: Conflict
      summary: Delete an account
      tags:
      - admin
    patch:
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        schema:
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/api.userUpdate'
              description: Settings
              summary: body
        description: Settings
        required: true
      responses:
        "204":
          description: No Content
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.apiError'
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.apiError'
          description: Not Found
      summary: Update an account. A new stream limit applies from the next title played.
      tags:
      - admin
...
  /auth/login:
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/api.credentials'
              description: Credentials
              summary: body
        description: Credentials
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/store.User'
          description: OK
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.apiError'
          description: Unauthorized
        "429":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.apiError'
          description: 'Too many failed attempts from this IP: retry after Retry-After
            seconds'
      summary: Log in
      tags:
      - auth
...
  /me/password:
    put:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/api.passwordChange'
              description: Passwords
              summary: body
        description: Passwords
        required: true
      responses:
        "204":
          description: No Content
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.apiError'
          description: Unauthorized
      summary: Change the account password (logs out every session)
      tags:
      - auth
...
  /profiles:
    get:
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/store.Profile'
                type: array
          description: OK
      summary: Profiles of the account
      tags:
      - profiles
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/api.profileBody'
              description: Profile
              summary: body
        description: Profile
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/store.Profile'
          description: Created
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.apiError'
          description: Conflict
      summary: Create a profile (5 max per account)
      tags:
      - profiles
  /profiles/{id}:
    delete:
      parameters:
      - description: Profile ID
        in: path
        name: id
        required: true
        schema:
          type: integer
      responses:
        "204":
          description: No Content
      summary: Delete a profile and its history
      tags:
      - profiles
    patch:
      parameters:
      - description: Profile ID
        in: path
        name: id
        required: true
        schema:
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/api.profileBody'
              description: Profile
              summary: body
        description: Profile
        required: true
      responses:
        "204":
          description: No Content
      summary: Rename or change the avatar of a profile
      tags:
      - profiles
  /setup:
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/api.setupRequest'
              description: Admin credentials and language
              summary: body
        description: Admin credentials and language
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/store.User'
          description: Created
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.apiError'
          description: Conflict
      summary: Create the first admin account and pick the instance language (only
        while no account exists), then log in
      tags:
      - auth
...
servers:
- url: /api/v1